Works with examples gives
Removed the arrays and worked only with the input string and values at each index. While loop runs while the current index being looked at is less than the length of the string input. If the current character at index i is a number look at value at (the current index + the number + 1) and append it to the string "decoded". If the current character is not a number move to the next index (I don't know if this is necessary if everything is coded the same way). Does not have inline comments